Updating gas *.pyx files to work with recent versions of Numpy
Adding `allowPickle=True` to the `np.load()` calls in all the gas files
to work with recent versions of Numpy, rather than having to downgrade.

Numpy past version 1.16.2 sets `allowPickle=False` as the default.

Also including a script to apply the change automatically to all gas
files for users not working from this repository.